# api/web_page/openapi.yaml
swagger: "2.0"
info:
    title: Mai.Today Server
    description: Mai.Today Server
    version: 1.0.0
host: dev.api.mai.today
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /{channel}/createBase:
        get:
            tags:
                - Base
            summary: '[WebSocket] 創造一個新的基地'
            operationId: Base#receiveCreateBase
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseReceiveCreateBaseOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /{channel}/deleteBase:
        get:
            tags:
                - Base
            summary: '[WebSocket] 删除指定的基地'
            operationId: Base#receiveDeleteBase
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseReceiveDeleteBaseOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /{channel}/reorderBaseNavStates:
        get:
            tags:
                - Base
            summary: '[WebSocket] 變更基地的順序'
            operationId: Base#receiveReorderBaseNavStates
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseReceiveReorderBaseNavStatesOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /{channel}/updateBaseInfo:
        get:
            tags:
                - Base
            summary: '[WebSocket] 更新基地資訊'
            operationId: Base#receiveUpdateBaseInfo
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseReceiveUpdateBaseInfoOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /base:
        post:
            tags:
                - Base
            summary: 創造一個新的基地
            operationId: Base#CreateBase
            parameters:
                - name: Authorization
                  in: header
                  required: false
                  type: string
                - name: CreateBaseRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/BaseCreateBaseRequestBody'
                    required:
                        - color
                        - logo
                        - name
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseCreateBaseOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "400":
                    description: 參數錯誤
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/BaseCreateBaseInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
    /base/{id}:
        delete:
            tags:
                - Base
            summary: 删除指定的基地
            operationId: Base#DeleteBase
            parameters:
                - name: id
                  in: path
                  description: 識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseDeleteBaseOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/BaseDeleteBaseInvalidTokenResponseBody'
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/BaseDeleteBasePermissionDeniedResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/BaseDeleteBaseNotFoundResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
    /base/{id}/info:
        put:
            tags:
                - Base
            summary: 更新基地資訊
            operationId: Base#UpdateBaseInfo
            parameters:
                - name: id
                  in: path
                  description: 識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
                - name: UpdateBaseInfoRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/BaseUpdateBaseInfoRequestBody'
                    required:
                        - color
                        - logo
                        - name
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseUpdateBaseInfoOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "400":
                    description: 參數錯誤
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/BaseUpdateBaseInfoInvalidTokenResponseBody'
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/BaseUpdateBaseInfoPermissionDeniedResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/BaseUpdateBaseInfoNotFoundResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
    /base/nav-state/reorder:
        put:
            tags:
                - Base
            summary: 變更基地的順序
            operationId: Base#ReorderBaseNavStates
            parameters:
                - name: Authorization
                  in: header
                  required: false
                  type: string
                - name: ReorderBaseNavStatesRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/BaseReorderBaseNavStatesRequestBody'
                    required:
                        - id
                        - newIndex
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/BaseReorderBaseNavStatesOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "400":
                    description: 參數錯誤
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/BaseReorderBaseNavStatesInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
definitions:
...
    BaseReceiveCreateBaseOKResponseBody:
        title: BaseReceiveCreateBaseOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/CreateBaseResultDataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: createBase
            data:
                id: 669071b8a650a662b82285ca
                info:
                    baseId: EbZvovBKbFogJeRem5oj2S
                    color: '#FF5733'
                    deleted_at: null
                    id: EbZvovBKbFogJeRem5oj2S
                    logo: http://example.com/logo.png
                    name: 基地 A
                navState:
                    baseId: 669071b8a650a662b82285ca
                    deleted_at: "2024-01-01T11:53:54Z"
                    id: 669071b8a650a662b82285ca
                    index: 0
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
    BaseReceiveDeleteBaseOKResponseBody:
        title: BaseReceiveDeleteBaseOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/DeleteBaseResultDataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: deleteBase
            data:
                baseId: 669071b8a650a662b82285ca
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
...
    BaseReceiveUpdateBaseInfoOKResponseBody:
        title: BaseReceiveUpdateBaseInfoOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/BaseInfoResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: updateBaseInfo
            data:
                baseId: EbZvovBKbFogJeRem5oj2S
                color: '#FF5733'
                deleted_at: null
                id: EbZvovBKbFogJeRem5oj2S
                logo: http://example.com/logo.png
                name: 基地 A
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
...
    BaseReorderBaseNavStatesRequestBody:
        title: BaseReorderBaseNavStatesRequestBody
        type: object
        properties:
            id:
                type: string
                description: 識別碼
                example: 669071b8a650a662b82285ca
                minLength: 24
                maxLength: 24
            newIndex:
                type: integer
                description: 新的順序
                default: 1
                example: 0
                format: int64
                minimum: 1
        example:
            id: 669071b8a650a662b82285ca
            newIndex: 0
        required:
            - id
            - newIndex
...
securityDefinitions:
    JWT_header_Authorization:
        type: apiKey
        name: Authorization
        in: header